feat: generate CSV file on the server using a server action
